# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ECOMMERCE API Documentación",
        "description": "ECOMMERCE API Documentación",
        "contact": {
            "email": "fericell2909@gmail.com"
        },
        "license": {
            "name": "GPL2",
            "url": "https://devsenv.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Gestión de Usuarios"
                ],
                "summary": "HU-000: Iniciar Sesión",
                "description": "HU-000: Iniciar Sesión",
                "operationId": "8502ae0150b1a339dccf1ee3bd965a72",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "fericell29091@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "abcd1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Acceso Correcto.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseLogin200Schema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitu Inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseLogin400Schema"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso Restringido.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseLogin403Schema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró el recurso en el servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseGeneric404Schema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Interno del Servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseLogin500Schema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Gestión de Usuarios"
                ],
                "summary": "HU-000 : Cerrar Sesión",
                "description": "Cerrar Sesión del Usuario Autenticado.",
                "operationId": "263f484d50d35eb12eebb198f799c85f",
                "responses": {
                    "200": {
                        "description": "Cierre de Sesión.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseLogout200Schema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acceso no Autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseGeneric401Schema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Interno del Servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseGeneric500Schema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ResponseGeneric401Schema": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Indica que no puede acceder a la solicitud ya que no esta autenticado.",
                        "type": "string",
                        "example": "Acceso no Autenticado."
                    },
                    "errors": {
                        "type": "null",
                        "example": null
                    },
                    "data": {
                        "type": "null",
                        "example": null
                    }
                },
                "type": "object"
            },
...
            "ResponseGeneric404Schema": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Indica que no se encontro la solicitud en el servidor.",
                        "type": "string",
                        "example": "La solicitud no ha sido encontrada en el servidor."
                    },
                    "errors": {
                        "type": "null",
                        "example": null
                    },
                    "data": {
                        "type": "null",
                        "example": null
                    }
                },
                "type": "object"
            },
            "ResponseGeneric500Schema": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Mensaje que se obtiene de getMessage de la Excepción.",
                        "type": "string",
                        "example": "Ha ocurrido un error en el servidor"
                    },
                    "errors": {
                        "type": "null"
                    },
                    "data": {
                        "type": "null"
                    },
                    "ex": {
                        "description": "Se transforma a string el objeto de la Excepción ocurrida en el servidor",
                        "type": "string",
                        "example": "..."
                    }
                },
                "type": "object"
            },
...
            "ResponseLogin200Schema": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Acceso Correcto"
                    },
                    "errors": {
                        "type": "null",
                        "example": "null"
                    },
                    "data": {
                        "properties": {
                            "token_type": {
                                "description": "Tipo de Token",
                                "type": "string",
                                "example": "Bearer"
                            },
                            "accessToken": {
                                "description": "Token de Acceso",
                                "type": "string",
                                "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9"
                            },
                            "expires_in": {
                                "description": "Valor Expresado en segundos",
                                "type": "integer",
                                "example": 3600
                            },
                            "id": {
                                "description": "ID usuario logueado.",
                                "type": "integer",
                                "example": 1
                            },
                            "email": {
                                "description": "Correo Electrónico de usuario autenticado",
                                "type": "string",
                                "example": "fericell2909@gmail.com"
                            },
                            "name": {
                                "description": "Nombres de usuario autenticado",
                                "type": "string",
                                "example": "Marco"
                            },
                            "surnames": {
                                "description": "Apellidos de usuario autenticado",
                                "type": "string",
                                "example": "Estrada López"
                            },
                            "language": {
                                "description": "Lenguaje de usuario autenticado",
                                "type": "string",
                                "example": "es"
                            },
                            "uroles": {
                                "description": "Roles asociados al usuario authenticado",
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "code": {
                                            "description": "Código de Rol",
                                            "type": "string",
                                            "example": "ST"
                                        },
                                        "name": {
                                            "description": "Nombre del Rol",
                                            "type": "string",
                                            "example": "Supervisor Tributario"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "roles": {
                                "description": "Listado de Roles que se encuentran en el sistema.",
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "code": {
                                            "description": "Código de Rol",
                                            "type": "string",
                                            "example": "AT"
                                        },
                                        "name": {
                                            "description": "Nombre del Rol",
                                            "type": "string",
                                            "example": "Analista Tributario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ResponseLogin400Schema": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Puede venir mensaje de Error accediendo a la cuenta o correo electrónico sin verificar.",
                        "type": "string",
                        "example": "Error accediendo a la cuenta."
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": [
                                        "El e-mail debe tener al menos 8 caracteres.",
                                        "El e-mail no debe ser mayor a 255 caracteres.",
                                        "El formato del e-mail no es válido.",
                                        "El e-mail es requerido."
                                    ]
                                }
                            },
                            "password": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": [
                                        "La contraseña es requerida."
                                    ]
                                }
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "null",
                        "example": "null"
                    }
                },
                "type": "object"
            },
            "ResponseLogin403Schema": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Puede venir mensaje de Error accediendo a la cuenta o correo electrónico sin verificar.",
                        "type": "string",
                        "example": "Error accediendo a la cuenta"
                    },
                    "errors": {
                        "type": "null",
                        "example": "null"
                    },
                    "data": {
                        "type": "null",
                        "example": "null"
                    }
                },
                "type": "object"
            },
            "ResponseLogin500Schema": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Mensaje que se obtiene de getMessage de la Excepción.",
                        "type": "string",
                        "example": "Ha ocurrido un error en el servidor"
                    },
                    "errors": {
                        "type": "null"
                    },
                    "data": {
                        "type": "null"
                    },
                    "ex": {
                        "description": "Se transforma a string el objeto de la Excepción ocurrida en el servidor",
                        "type": "string",
                        "example": "..."
                    }
                },
                "type": "object"
            },
            "ResponseLogout200Schema": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Mensaje de respuesta",
                        "type": "string",
                        "example": "Cierre sesión exitosamente"
                    },
                    "errors": {
                        "type": "null",
                        "example": "null"
                    },
                    "data": {
                        "type": "null",
                        "example": "null"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}